Implementação da inicialização do banco de dados no start da aplicação.

+++ Api/ViewModels/TaxaViewModel.cs
@@ -5,6 +5,6 @@ namespace Api.ViewModel
     public class TaxaViewModel
     {
         public Guid Id { get; set; }
-        public decimal Taxa { get; set; }
+        public double Taxa { get; set; }
     }
 }
